refactor: optimize logging system and fix compiler warnings
- Add DebugLoggerProvider for VS Output window in Debug builds
- Extract logging configuration constants in LumiereLoggerFactory
- Remove ValidationLogger (dead code, only bridge call remained)
- Fix CS0675 bitwise-or sign extension warning
- Remove unused hasKeyboardFocus field (CS0414)

+++ src/Lumiere.App/TrayMenuWindow.xaml.cs
@@ -269,7 +269,7 @@ public sealed partial class TrayMenuWindow : Window
     private static void AddWindowExStyle(IntPtr hWnd, int exStyle)
     {
         var current = GetWindowLongPtr(hWnd, GwlExstyle);
-        SetWindowLongPtr(hWnd, GwlExstyle, new IntPtr(current.ToInt64() | exStyle));
+        SetWindowLongPtr(hWnd, GwlExstyle, new IntPtr(current.ToInt64() | (uint)exStyle));
     }
 
     private void BringToTopmost(int x, int y, int width, int height)
